TICTACTOE
This project presents a refined implementation of the classic Tic Tac Toe game, developed using HTML, CSS, and JavaScript. The interface is designed for a seamless user experience, offering responsive functionality and a clean layout. Engage in a strategic challenge with a friend or enjoy solo gameplay with this timeless classic.